macOS fixes (#998)
* Fix creation of slade.pk3 on macOS

Charset command line option is not supported on macOS

* Fix compatibility issue with wxWidgets 3.1.0

src/TextEditor/Lexer.cpp:119:28: error: too few arguments to function call, expected 2, have 1

wxStyledTextCtrl::StartStyling() has one parameter removed in wxWidgets 3.1.1
https://github.com/wxWidgets/wxWidgets/blob/v3.1.0/include/wx/stc/stc.h
https://github.com/wxWidgets/wxWidgets/blob/v3.1.1/include/wx/stc/stc.h
https://github.com/wxWidgets/wxWidgets/blob/v3.1.2/include/wx/stc/stc.h

* Fix initializer lists compilation issues with Clang

src/Dialogs/ModifyOffsetsDialog.cpp:121:11: error: non-constant-expression cannot be narrowed from type 'long' to 'int' in initializer list [-Wc++11-narrowing]
src/Dialogs/ModifyOffsetsDialog.cpp:121:14: error: non-constant-expression cannot be narrowed from type 'long' to 'int' in initializer list [-Wc++11-narrowing]
src/MapEditor/Renderer/Overlays/LineInfoOverlay.cpp:61:48: error: type 'double' cannot be narrowed to 'int' in initializer list [-Wc++11-narrowing]
src/MapEditor/Renderer/Overlays/ThingInfoOverlay.cpp:63:86: error: type 'double' cannot be narrowed to 'int' in initializer list [-Wc++11-narrowing]

* Fix compilation of RunDialog on macOS

src/Dialogs/RunDialog.cpp:566:9: error: use of undeclared identifier 'text_exe_path'

+++ src/MapEditor/Renderer/Overlays/LineInfoOverlay.cpp
@@ -58,7 +58,7 @@
 // -----------------------------------------------------------------------------
 LineInfoOverlay::LineInfoOverlay() :
 	scale_{ Drawing::fontSize() / 12.0 },
-	text_box_{ "", Drawing::Font::Condensed, 100, 16 * scale_ }
+	text_box_{ "", Drawing::Font::Condensed, 100, int(16 * scale_) }
 {
 }
